+++ agent/api/task_unix.go
@@ -20,6 +20,7 @@ import (
 	"time"
 
 	"github.com/aws/amazon-ecs-agent/agent/config"
+	"github.com/cihub/seelog"
 	docker "github.com/fsouza/go-dockerclient"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
@@ -39,6 +40,9 @@ const (
 	minimumCPUPercent = 0
 	bytesPerMegabyte  = 1024 * 1024
 )
+
+// platformFields consists of fields specific to Linux for a task
+type platformFields struct {}
 
 func (task *Task) adjustForPlatform(cfg *config.Config) {
 	task.lock.Lock()
@@ -177,4 +181,18 @@ func (task *Task) overrideCgroupParent(hostConfig *docker.HostConfig) error {
 		hostConfig.CgroupParent = cgroupRoot
 	}
 	return nil
+}
+
+// dockerCPUShares converts containerCPU shares if needed as per the logic stated below:
+// Docker silently converts 0 to 1024 CPU shares, which is probably not what we
+// want.  Instead, we convert 0 to 2 to be closer to expected behavior. The
+// reason for 2 over 1 is that 1 is an invalid value (Linux's choice, not Docker's).
+func (task *Task) dockerCPUShares(containerCPU uint) int64 {
+	if containerCPU <= 1 {
+		seelog.Debugf(
+			"Converting CPU shares to allowed minimum of 2 for task arn: [%s] and cpu shares: %d",
+			task.Arn, containerCPU)
+		return 2
+	}
+	return int64(containerCPU)
 }